api-raspberry-pines/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
RPi.GPIO==0.7.0